int -> float

+++ internal/model/record.go
@@ -7,12 +7,12 @@ type Record struct {
 	ExcavatorName  string    `gorm:"not null" json:"excavator_name"`
 	Date           time.Time `gorm:"not null" json:"date"`
 	Shift          string    `gorm:"not null" json:"shift"`
-	ShiftTime      int       `gorm:"not null" json:"shift_time"`
+	ShiftTime      float64       `gorm:"not null" json:"shift_time"`
 	LoadTime       float64   `gorm:"not null" json:"load_time"`
-	CycleTime      int       `gorm:"not null" json:"cycle_time"`
-	ApproachTime   int       `gorm:"not null" json:"approach_time"`
+	CycleTime      float64       `gorm:"not null" json:"cycle_time"`
+	ApproachTime   float64       `gorm:"not null" json:"approach_time"`
 	ActualTrucks   float64   `gorm:"not null" json:"actual_trucks"`
-	Productivity   int       `gorm:"not null" json:"productivity"`
+	Productivity   float64       `gorm:"not null" json:"productivity"`
 	RequiredTrucks float64   `gorm:"not null" json:"required_trucks"`
 	PlanVolume     float64   `gorm:"not null" json:"plan_volume"`
 	ForecastVolume float64   `gorm:"not null" json:"forecast_volume"`
@@ -25,12 +25,12 @@ type RecordForAnalise struct {
 	ExcavatorName  string    `json:"excavator_name"`
 	Date           time.Time `json:"date"`
 	Shift          string    `json:"shift"`
-	ShiftTime      int       `json:"shift_time"`
+	ShiftTime      float64       `json:"shift_time"`
 	LoadTime       float64   `json:"load_time"`
-	CycleTime      int       `json:"cycle_time"`
-	ApproachTime   int       `json:"approach_time"`
+	CycleTime      float64       `json:"cycle_time"`
+	ApproachTime   float64       `json:"approach_time"`
 	ActualTrucks   float64   `json:"actual_trucks"`
-	Productivity   int       `json:"productivity"`
+	Productivity   float64       `json:"productivity"`
 	RequiredTrucks float64   `json:"required_trucks"`
 	PlanVolume     float64   `json:"plan_volume"`
 	ForecastVolume float64   `json:"forecast_volume"`
